refactoring

+++ autohdl/xilinx.py
@@ -3,12 +3,10 @@ import shutil
 import subprocess
 import sys
 import time
+from time import strftime
 
 from autohdl import toolchain
-from autohdl.hdlGlobals import implementPath
 from autohdl.hdlLogger import log_call, logging
-from autohdl import git
-from time import strftime
 
 
 log = logging.getLogger(__name__)
@@ -33,45 +31,42 @@ def bit2mcs(config):
         sys.exit()
 
 
-@log_call
-def clean(apath):
-    if os.path.exists(apath):
-        for root, dirs, files in os.walk(apath):
-            for f in files:
-                if os.path.splitext(f)[1] not in ['.mcs', '.bit']:
-                    file_to_del = os.path.join(root, f)
-                    try:
-                        os.remove(file_to_del)
-                    except OSError:
-                        log.warning("Can't delete file " + file_to_del)
-    if not os.path.exists(apath):
-        os.makedirs(apath)
-
-
-# TODO: that code should be in other place (git)
 def copy_firmware(config):
     firmware_ext = ['.bit', '.mcs']
-    dest_dir = os.path.join(config['hdlManager']['dsn_root'], 'resource')
-    for i in os.listdir(dest_dir):
-        if os.path.splitext(i)[1] in firmware_ext:
-            log.info('removing ' + os.path.join(dest_dir, i))
-            os.remove(os.path.join(dest_dir, i))
+    build_timestamp = strftime('%y%m%d_%H%M%S', time.localtime())
+    dest_dirs = [os.path.join(config['hdlManager']['dsn_root'], 'resource'),
+                 os.path.join(config['hdlManager']['dsn_root'], 'autohdl', 'firmware')]
+    for dest_dir in dest_dirs:
+        if not os.path.exists(dest_dir):
+            os.makedirs(dest_dir)
+        dest = '{dsn}_{top}_build_'.format(
+            dsn=os.path.basename(os.path.abspath('..')),
+            top=config['hdlManager'].get('top'))
+        for i in os.listdir(dest_dir):
+            if os.path.splitext(i)[1] in firmware_ext and dest in i:
+                log.info('removing ' + os.path.join(dest_dir, i))
+                try:
+                    os.remove(os.path.join(dest_dir, i))
+                except Exception as e:
+                    log.warning(e)
 
-    for i in firmware_ext:
-        src_abs_path = os.path.join(config['hdlManager']['dsn_root'],
-                                    'autohdl/implement',
-                                    config['hdlManager'].get('top'),
-                                    config['hdlManager'].get('top') + i)
-        if os.path.exists(src_abs_path):
-            dest = '{dsn}_{top}_build_{ver}{ext}'.format(
-                dsn=os.path.basename(os.path.abspath('..')),
-                top=config['hdlManager'].get('top'),
-                ver=strftime('%y%m%d_%H%M%S', time.localtime()),
-                ext=i
-            )
-            dest_abs_path = os.path.join(dest_dir, dest)
-            log.info('copying ' + dest_abs_path)
-            shutil.copy(src_abs_path, dest_abs_path)
+        for i in firmware_ext:
+            src_abs_path = os.path.join(config['hdlManager']['dsn_root'],
+                                        'autohdl/implement',
+                                        config['hdlManager'].get('top'),
+                                        config['hdlManager'].get('top') + i)
+            if os.path.exists(src_abs_path):
+                dest = '{dsn}_{top}_build_{ver}{ext}'.format(
+                    dsn=os.path.basename(os.path.abspath('..')),
+                    top=config['hdlManager'].get('top'),
+                    ver=build_timestamp,
+                    ext=i
+                )
+                dest_abs_path = os.path.join(dest_dir, dest)
+                log.info('copying ' + dest_abs_path)
+                shutil.copy(src_abs_path, dest_abs_path)
+
+
 
 
 xtcl_script = """
@@ -84,7 +79,7 @@ project set speed "{speed}"
 
 project set top_level_module_type "EDIF"
 project set "Preferred Language" "Verilog"
-
+project set "Verilog Macros" "I_HAVE_ONLY_FUCKING_XST"
 #project set "FPGA Start-Up Clock" "JTAG Clock"
 
 {xfiles}
@@ -161,8 +156,17 @@ def run_project(config):
     # check project.tcl src files if they are not modified, don't clean
     clean_project(config)
 
-    if not os.path.exists(config['xilinx']['project_directory']):
-        os.makedirs(config['xilinx']['project_directory'])
+    for i in range(3):
+        if not os.path.exists(config['xilinx']['project_directory']):
+            try:
+                os.makedirs(config['xilinx']['project_directory'])
+                break
+            except Exception as e:
+                log.warning(e)
+                time.sleep(1)
+        if not os.path.exists(config['xilinx']['project_directory']):
+            sys.exit("Cant create xilinx project")
+
     config['xilinx']['script_path'] = os.path.abspath(os.path.join('..',
                                                                    'autohdl',
                                                                    'implement',
@@ -179,37 +183,43 @@ def form_project_src(config):
     """
      return a list of source files as absolute paths
     """
-    synthesis_result_netlist = os.path.abspath(os.path.join('..',
-                                                            'autohdl',
-                                                            'synthesis',
-                                                            config['hdlManager']['top'],
-                                                            config['hdlManager']['top'] + '.edf', ))
-    log.info('Searching a netlist after synplify')
-    for i in range(30):
-        print "."
-        if os.path.exists(synthesis_result_netlist):
-            break
-        time.sleep(0.5)
-    if not os.path.exists(synthesis_result_netlist):
-        log.error('Cannot find netlist. Try synthesis first.')
-        sys.exit(1)
-
-    synthesis_result_constraint = os.path.abspath(os.path.join('..',
-                                                               'autohdl',
-                                                               'synthesis',
-                                                               config['hdlManager']['top'],
-                                                               'synplicity.ucf', ))
-    if not os.path.exists(synthesis_result_constraint):
-        print 'cant find', synthesis_result_constraint
-        synthesis_result_constraint = config['hdlManager']['ucf']
-    if not os.path.exists(synthesis_result_constraint):
-        log.warning('No constraint file')
-    netlists = config['structure'].get('netlists') or []  # should type list
     impl_src = set()
+    if True: #config.get('synthesis') == 'synplify':
+        synthesis_result_netlist = os.path.abspath(os.path.join('..',
+                                                                'autohdl',
+                                                                'synthesis',
+                                                                config['hdlManager']['top'],
+                                                                config['hdlManager']['top'] + '.edf', ))
+        log.info('Searching a netlist after synplify')
+        for i in range(30):
+            print "."
+            if os.path.exists(synthesis_result_netlist):
+                break
+            time.sleep(0.5)
+        if not os.path.exists(synthesis_result_netlist):
+            log.error('Cannot find netlist. Try synthesis first.')
+            sys.exit(1)
+
+        synthesis_result_constraint = os.path.abspath(os.path.join('..',
+                                                                   'autohdl',
+                                                                   'synthesis',
+                                                                   config['hdlManager']['top'],
+                                                                   'synplicity.ucf', ))
+        impl_src.add(synthesis_result_netlist)
+        if not os.path.exists(synthesis_result_constraint):
+            print 'cant find', synthesis_result_constraint
+            synthesis_result_constraint = config['hdlManager']['ucf']
+        if not os.path.exists(synthesis_result_constraint):
+            log.warning('No constraint file')
+    else:
+        synthesis_result_constraint = config['hdlManager']['ucf']
+        for i in config['structure']['mainSrc']+config['structure']['depSrc']:
+            impl_src.add(i)
     impl_src.add(synthesis_result_constraint)
-    impl_src.add(synthesis_result_netlist)
+    netlists = config['structure'].get('netlists') or []  # should type list
     for i in netlists:
         impl_src.add(i)
+
 
     config['xilinx']['src'] = impl_src
     return impl_src
